requirements_versions.txt
torchsde==0.2.5
einops==0.4.1
transformers==4.30.2
safetensors==0.3.1
accelerate==0.21.0
pyyaml==6.0
Pillow==9.2.0
scipy==1.9.3
tqdm==4.64.1
psutil==5.9.5
pytorch_lightning==1.9.4
omegaconf==2.2.3
gradio==3.41.2
pygit2==1.12.2
opencv-contrib-python==4.8.0.74
httpx==0.24.1
onnxruntime==1.16.3
timm==0.9.2

protobuf>=4.23.2
addict
future
lmdb
numpy
opencv-python
requests
scikit-image
tb-nightly
yapf
lpips
insightface>=0.7.3

torch==2.0.1
torchvision==0.15.2
diffusers==0.25.0
huggingface-hub==0.20.2
spaces==0.19.4
safetensors
omegaconf
peft